AZ#525: Merge duplicate showReport page files
